testing 'findSequenceSegments' plus a small bug fix
